Separate game menu from studio
