+++ src/studio/screens/game_menu.h
@@ -22,25 +22,7 @@
 
 #pragma once
 
-#include "../system.h"
-#include "studio/studio.h"
-
-typedef struct Run Run;
-
-struct Run
-{
-    tic_mem* tic;
-    struct Console* console;
-    struct tic_fs* fs;
-    tic_tick_data tickData;
-
-    bool exit;
-    
-    char saveid[TICNAME_MAX];
-    tic_persistent pmem;
-
-    void(*tick)(Run*);
-};
-
-void initRun(Run*, struct Console*, struct tic_fs* , tic_mem*);
-void freeRun(Run* run);
+void showGamepadMenu();
+void showMainMenu();
+void freeGameMenu();
+void initGamepadMenu();
